requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0

# Video and Audio Processing
opencv-python>=4.5.0
librosa>=0.9.0
moviepy>=1.0.0
imageio>=2.9.0

# Machine Learning and Vision
jax>=0.3.0
jaxlib>=0.3.0
flax>=0.4.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0

# VideoPrism (Video Embeddings)
videoprism>=0.1.0

# Whisper (Audio Transcription)
openai-whisper>=20230314

# LLM Integration
openai>=1.0.0

# Vector Search
faiss-cpu>=1.7.0

# CLI
typer>=0.4.0
click>=8.0.0

# Utilities
tqdm>=4.62.0
pydantic>=1.9.0
pyyaml>=5.4.0

# Development (optional)
pytest>=6.2.0
black>=21.0
flake8>=3.9.0
